refactor: (#81) 네이밍 수정 PostDay -> DayOfWeek
