+++ src/main/kotlin/org/example/root_be/domain/day_of_week/domain/repository/DayOfWeekRepository.kt
@@ -1,7 +1,7 @@
-package org.example.root_be.domain.post_day.domain.repository
+package org.example.root_be.domain.day_of_week.domain.repository
 
 import org.example.root_be.domain.post.domain.VolunteerPost
-import org.example.root_be.domain.post_day.domain.DayOfWeek
+import org.example.root_be.domain.day_of_week.domain.DayOfWeek
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface DayOfWeekRepository: JpaRepository<DayOfWeek, Long> {
