+++ src/main/kotlin/org/example/root_be/domain/post/service/ModifyVolunteerPostService.kt
@@ -7,9 +7,9 @@ import org.example.root_be.domain.post.domain.VolunteerPost
 import org.example.root_be.domain.post.domain.repository.VolunteerPostRepository
 import org.example.root_be.domain.post.facade.VolunteerFacade
 import org.example.root_be.domain.post.presentation.dto.request.ModifyVolunteerPostRequest
-import org.example.root_be.domain.post_day.domain.DayOfWeek
-import org.example.root_be.domain.post_day.domain.repository.DayOfWeekRepository
-import org.example.root_be.domain.post_day.exception.DayOfWeekNotFoundException
+import org.example.root_be.domain.day_of_week.domain.DayOfWeek
+import org.example.root_be.domain.day_of_week.domain.repository.DayOfWeekRepository
+import org.example.root_be.domain.day_of_week.exception.DayOfWeekNotFoundException
 import org.example.root_be.domain.role.domain.VolunteerRole
 import org.example.root_be.domain.role.domain.repository.RoleRepository
 import org.example.root_be.domain.role.exception.VolunteerRoleNotFoundException
